optimize code API test
