+++ app/src/test/java/vn/asiantech/way/api/ApiSuiteTest.kt
@@ -15,7 +15,9 @@ import org.junit.runners.Suite
         GoogleApiTest::class,
         GetUserInfoTest::class,
         GetMemberListTest::class,
-        AddUserToGroupTest::class
+        AddUserToGroupTest::class,
+        GetGroupInfoTest::class,
+        CreateGroupTest::class
 )
 class ApiSuiteTest {
     companion object {
